Upgrade to PHPUnit 9
We have to finally replace the `::assertTag()` verification, and use
uopz instead of runkit due to better availability.  We completely
disable the ControllerTest for now; this is likely going to be dropped
anyway.

It's interesting to see that I reached a pretty good code coverage
already more than ten years ago, although the test setup was (and is)
super complex.